# api/openapi.json
{
  "info": {
    "title": "Aldus API",
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/v1/{chain}/{network}/entities": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "chain",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "network",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Entity"
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "V1.Entities"
        ],
        "summary": "Get entities with details",
        "description": "Get all entities entry and details"
      }
    },
...
    "/v1/{chain}/{network}/entities/{entity_slug}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "chain",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "network",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "entity_slug",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "V1.Entities"
        ],
        "summary": "Get entity details by slug",
        "description": "Get entity entry and details by slug"
      }
    },
...
  },
  "openapi": "3.0.3",
  "components": {
    "schemas": {
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Social": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ]
      },
      "Account": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "address",
          "description",
          "name",
          "slug",
          "type"
        ]
      },
      "Code": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "github": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "github",
          "id",
          "name",
          "slug"
        ]
      },
      "Contract": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "github": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "github",
          "id",
          "name",
          "slug"
        ]
      },
      "Module": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "github": {
            "type": "string"
          }
        },
        "required": [
          "address",
          "description",
          "github",
          "name",
          "slug"
        ]
      },
...
      "EntityDetail": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "github": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "socials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Social"
            }
          },
          "website": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "github",
          "logo",
          "name",
          "socials",
          "website"
        ]
      },
      "Entity": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/EntityDetail"
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "codes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Code"
            }
          },
          "contracts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contract"
            }
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          }
        },
        "required": [
          "details",
          "slug"
        ]
      }
    }
  }
}